refactor(Banks Dashboard): factorized code
