+++ packages/cozy-konnector-libs/src/libs/categorization/dashboard/services-global-fixtures.spec.js
@@ -5,6 +5,11 @@ const cat2name = require('./tree.json')
 const allowedFallbackCategories = require('./allowed_wrong_categories.json')
 const { fetchTransactionsWithManualCat } = require('../localModel/parameters')
 const { fetchParameters } = require('../globalModel/parameters')
+const {
+  softRequire,
+  fmtManualCategorizations,
+  fmtResultsCSV
+} = require('./helpers')
 
 jest.mock('../localModel/parameters')
 jest.mock('../globalModel/parameters')
@@ -13,14 +18,6 @@ const fixturePath = path.join(__dirname, 'fixtures')
 
 const BACKUP_DIR = process.env.BACKUP_DIR
 const IT_IS_A_TEST = process.env.IT_IS_A_TEST
-
-const softRequire = file => {
-  try {
-    return require(file)
-  } catch (e) {
-    return undefined
-  }
-}
 
 const globalModelJSON = softRequire('./bank_classifier_nb_and_voc.json')
 const xOrDescribe = globalModelJSON ? describe : xdescribe
@@ -128,48 +125,6 @@ const checkCategorization = transactions => {
   })
 }
 
-const compare = (a, b) => {
-  if (a.label < b.label) return -1
-  if (a.label > b.label) return 1
-  return 0
-}
-
-const fmtManualCategorizations = manualCategorizations => {
-  const sortedManualCategorizations = manualCategorizations.sort(compare)
-  let countOfManualCategorizations = {}
-  // sum up every recategorizations
-  for (const op of sortedManualCategorizations) {
-    const key =
-      op.label.slice(0, 15) + op.automaticCategoryId + '>' + op.manualCategoryId
-    const operationsSummary = countOfManualCategorizations[key]
-    if (operationsSummary) {
-      countOfManualCategorizations[key] = {
-        occurrence: operationsSummary.occurrence + 1,
-        ...op
-      }
-    } else {
-      countOfManualCategorizations[key] = { occurrence: 1, ...op }
-    }
-  }
-  // display the summary
-  let fmtedManualCategorizations = []
-  for (const key of Object.keys(countOfManualCategorizations)) {
-    const op = countOfManualCategorizations[key]
-    const label = op.label
-    const manualCategoryName = cat2name[op.manualCategoryId]
-    const automaticCategoryName = cat2name[op.automaticCategoryId]
-    const formatedStr = `\t${
-      op.occurrence
-    } x <<${label}>>\t mapped from ${automaticCategoryName} to ${manualCategoryName}`
-    fmtedManualCategorizations.push(formatedStr)
-  }
-  const headOfSummary = [
-    `${manualCategorizations.length} Manual categorization for this fixture`
-  ]
-  const summary = headOfSummary.concat(fmtedManualCategorizations)
-  return summary
-}
-
 const fmtAccuracy = accuracy => {
   const {
     nOperations,
@@ -265,24 +220,6 @@ const fmtResults = transactions => {
   return fmtedResults
 }
 
-const fmtResultsCSV = (transactions, cozyId) => {
-  const fmtedResults = transactions.map(op => {
-    const { status, method, amount, label, catNameDisplayed, catNameTrue } = op
-    let fmtedResult = {
-      manCat: op.manualCategoryId !== undefined,
-      method,
-      status,
-      amount,
-      label,
-      catNameDisplayed,
-      catNameTrue,
-      cozyId
-    }
-    return fmtedResult
-  })
-  return fmtedResults
-}
-
 const computeAccuracy = transactions => {
   const nOperations = transactions.length
   let winGlobalModel = 0
@@ -334,7 +271,6 @@ xOrDescribe('Chain of predictions', () => {
     fetchTransactionsWithManualCat.mockImplementation(() =>
       Promise.resolve(manualCategorizations)
     )
-
     // Mock global model
     fetchParameters.mockImplementation(() => Promise.resolve(globalModelJSON))
   })
